@@ -6,27 +6,26 @@
     "@babel/core": "^7.4.3",
     "@babel/preset-env": "^7.4.3",
     "@babel/register": "^7.4.0",
-    "@signalk/client": "^2.0.4",
+    "@signalk/client": "^2.2.0",
     "axios": "^0.18.0",
     "konva": "^3.2.4",
     "luxon": "^1.12.1",
     "object.pick": "^1.3.0",
-    "react": "^16.8.6",
-    "react-dom": "^16.8.6",
-    "react-konva": "^16.8.6",
-    "react-redux": "^7.0.2",
+    "react": "^18.1.0",
+    "react-dom": "^18.1.0",
+    "react-konva": "^18.1.1",
+    "react-redux": "^8.0.2",
     "react-scripts": "2.1.8",
-    "redux": "^4.0.1",
+    "redux": "^4.2.0",
     "redux-devtools-extension": "^2.13.8",
-    "redux-saga": "^1.0.2",
-    "redux-thunk": "^2.3.0",
-    "typescript": "^3.4.3"
+    "redux-saga": "^1.1.3",
+    "redux-thunk": "^2.4.1"
   },
   "scripts": {
     "start": "react-scripts start",
     "build": "react-scripts build"
   },
-  "license": "Apache_2.0",
+  "license": "Apache-2.0",
   "eslintConfig": {
     "extends": "react-app"
   },
@@ -39,10 +38,11 @@
   "devDependencies": {
     "@types/jest": "^24.0.11",
     "@types/luxon": "^1.12.0",
-    "@types/node": "^11.13.4",
-    "@types/react": "^16.8.13",
-    "@types/react-dom": "^16.8.4",
-    "@types/react-redux": "^7.0.6",
-    "@types/redux-saga": "^0.10.5"
+    "@types/node": "^17.0.39",
+    "@types/react": "^18.0.11",
+    "@types/react-dom": "^18.0.5",
+    "@types/react-redux": "^7.1.24",
+    "@types/redux-saga": "^0.10.5",
+    "typescript": "^4.7.3"
   }
 }
